Add batched pipeline
