+++ wormhole-infra/wormhole-context/src/main/java/com/zergclan/wormhole/pipeline/filter/NotBlankValidator.java
@@ -15,11 +15,12 @@
  * limitations under the License.
  */
 
-package com.zergclan.wormhole.pipeline.filter.validator;
+package com.zergclan.wormhole.pipeline.filter;
 
 import com.zergclan.wormhole.api.Filter;
 import com.zergclan.wormhole.core.data.DataGroup;
 import com.zergclan.wormhole.core.data.DataNode;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 /**
@@ -27,6 +28,9 @@ import lombok.RequiredArgsConstructor;
  */
 @RequiredArgsConstructor
 public final class NotBlankValidator implements Filter<DataGroup> {
+    
+    @Getter
+    private final int order;
     
     private final String[] names;
     
@@ -41,5 +45,10 @@ public final class NotBlankValidator implements Filter<DataGroup> {
             }
         }
         return true;
+    }
+    
+    @Override
+    public String getType() {
+        return "NOT_BLANK_VALIDATOR";
     }
 }
