+++ wormhole-infra/wormhole-core/src/main/java/com/zergclan/wormhole/api/Filter.java
@@ -31,4 +31,18 @@ public interface Filter<D> {
      * @return Is success
      */
     boolean doFilter(D data);
+    
+    /**
+     * Get order.
+     *
+     * @return order
+     */
+    int getOrder();
+    
+    /**
+     * Get type.
+     *
+     * @return type
+     */
+    String getType();
 }
